Adding ZIO_FLAG_DIO_READ
The previous commit changed a Direct I/O read into an ARC read when
ECKSUM is returned. This can be cause by the data buffer for read being
manipulated while the Direct I/O read is occuring. However, with just
that patch alone, zpool status -v was still reporting checksum errors.
This new flag allows for the checksum errors not be logged at all for
Direct I/O reads, as those will just be redirected through the ARC. In
this case the ARC read will determine if there truly is data corruption
that took place.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>