Making Direct I/O completely stable for FreeBSD
Previously on Direct I/O writes were made stable to make sure the
buffere contents could not be manipulated while the I/O was in flight on
FreeBSD. This update now does the same thing for Direct I/O reads as
well. This allows FreeBSD to avoid having to use the ZIO_FLAG_DIO_READ
to disble self healing and reissuing through the ARC for Direct I/O
reads as is done on Linux.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>