Making Direct I/O completely stable for FreeBSD
Previously on Direct I/O writes were made stable to make sure the
buffere contents could not be manipulated while the I/O was in flight on
FreeBSD. This update now does the same thing for Direct I/O reads as
well. This allows FreeBSD to avoid having to use the ZIO_FLAG_DIO_READ
to disble self healing and reissuing through the ARC for Direct I/O
reads as is done on Linux. This update includes the following:

1. Returned FreeBSD's abd_return_buf() back to it's original
   implementation.
2. Updated comments in various parts of the code.
3. Updated man pages.
4. Created new ZTS test case dio_read_stable_pages.ksh for FreeBSD.
5. Made ZTS test dio_read_verify to only run for Linux.
6. Have all Direct I/O requests in FreeBSD call
   zfs_uio_set_pages_to_stable() and zfs_uio_release_stable_pages()
7. Updated zfs_uio_free_dio_pages() to longer require the UIO type.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>