This just needs to be squashed
However, I updated the ASSERT in zfs_uio_pin_user_pages() as nr_segs in
the iov_iter for a ITER_UBUF was not set until v6.4. We can just ASSERT
that the amount wanted, len, is >= 0. Also, we can short circuit trying
to pin the pages if len is 0.

Also, I updated the call to iov_iter_get_pages() as the last parameter
actually sets the start variable. It does not at all take into account
the skip amount that was being passed previously.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>